move metrics to a new server
