+++ pkg/metrics/metrics_test.go
@@ -1,10 +1,13 @@
 package metrics
 
 import (
+	"github.com/labstack/echo/v4"
 	promtestutil "github.com/prometheus/client_golang/prometheus/testutil"
 	clientmodel "github.com/prometheus/client_model/go"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"net/http"
+	"net/http/httptest"
 	"strings"
 	"testing"
 	"time"
@@ -144,4 +147,31 @@ func getExpectedLabelPairs() ([]clientmodel.LabelPair, []clientmodel.LabelPair, 
 		createLabelPairs("status_code", "500"),
 	}
 	return getSuccess, getFailure, listSuccess, listFailure
+}
+
+func TestMetricsHandler(t *testing.T) {
+	// Create a request to pass to our handler. We don't have any query parameters, so we'll
+	// pass 'nil' as the third parameter.
+	req, err := http.NewRequest(http.MethodGet, "/metrics", nil)
+	require.NoError(t, err)
+
+	// Create handler instance.
+	metricsCtrl := NewMetrics()
+	RegisterCustomMetrics()
+
+	t.Run("valid metrics json", func(t *testing.T) {
+		// We create a ResponseRecorder (which satisfies http.ResponseWriter) to record the response.
+		e := echo.New()
+		rec := httptest.NewRecorder()
+		ctx := e.NewContext(req, rec)
+
+		//when
+		err := metricsCtrl.PrometheusHandler(ctx)
+		require.NoError(t, err)
+		// then
+		// check the status code is what we expect.
+		require.Equal(t, http.StatusOK, rec.Code)
+		// check response content-type.
+		require.Equal(t, "text/plain; version=0.0.4; charset=utf-8", rec.Header().Get("Content-Type"))
+	})
 }
